Fixed bugs in HAWQ compatibility code
Also fixed a bug in madlib.glm's print function. This bug-fix needs to
be merged into master.

+++ R/madlib-glm.R
@@ -82,25 +82,27 @@ madlib.glm <- function (formula, data, family = "gaussian",
     ## construct SQL string
     conn.id <- conn.id(data)
     tbl.source <- gsub("\"", "", content(data))
-    tbl.output <- .unique.string()
     madlib <- schema.madlib(conn.id) # MADlib schema name
     if (db.str == "HAWQ") {
+        tbl.output <- NULL
         sql <- paste0("select (f).* from (select ", madlib,
                       ".logregr('", tbl.source, "', '", params$dep.str,
                       "', '", params$ind.str, "', ", max_iter,
                       ", '", method, "', ", tolerance, ") as f) s")
     } else {
+        tbl.output <- .unique.string()
         sql <- paste0("select ", madlib, ".logregr_train('",
                       tbl.source, "', '", tbl.output, "', '",
                       params$dep.str, "', '", params$ind.str, "', ",
                       grp, ", ", max_iter, ", '", method, "', ",
                       tolerance, ")")
+    }
 
     ## execute the logistic regression and get the result
     res <- .get.res(sql, tbl.output, conn.id)
 
     ## drop temporary tables
-    .db.removeTable(tbl.output, conn.id)
+    if (!is.null(tbl.output)) .db.removeTable(tbl.output, conn.id)
     if (is.tbl.source.temp) .db.removeTable(tbl.source, conn.id)
 
     msg.level <- .set.msg.level(msg.level, conn.id) # reset message level
@@ -135,7 +137,7 @@ madlib.glm <- function (formula, data, family = "gaussian",
         rst[[i]]$std_err <- r.std_err[i,]
         rst[[i]]$z_stats <- r.z_stats[i,]
         rst[[i]]$p_values <- r.p_values[i,]
-        rst[[i]]$odds_ratios <- r.odds_ratios
+        rst[[i]]$odds_ratios <- r.odds_ratios[i,]
         rst[[i]]$grp.cols <- r.grp.cols
         rst[[i]]$has.intercept <- r.has.intercept
         rst[[i]]$ind.vars <- r.ind.vars
@@ -263,10 +265,10 @@ print.logregr.madlib <- function (x,
     if (x$has.intercept)
         rows <- c("(Intercept)", x$ind.vars)
     else
-        rows <- x[[1]]$ind.vars
-    for (i in seq_len(length(x[[1]]$col.name)))
-        if (x[[1]]$col.name[i] != x[[1]]$appear[i])
-            rows <- gsub(x[[1]]$col.name[i], x[[1]]$appear[i], rows)
+        rows <- x$ind.vars
+    for (i in seq_len(length(x$col.name)))
+        if (x$col.name[i] != x$appear[i])
+            rows <- gsub(x$col.name[i], x$appear[i], rows)
     ind.width <- .max.width(rows)
 
     cat("\nMADlib Logistic Regression Result\n")
